Move pitch mss step file
[MAILPOET-2142]
