+++ assets/js/src/wizard/welcome_wizard_controller.jsx
@@ -7,7 +7,7 @@ import WelcomeWizardMigratedUserStep from './steps/migrated_user_step.jsx';
 import WelcomeWizardEmailCourseStep from './steps/email_course_step.jsx';
 import WelcomeWizardUsageTrackingStep from './steps/usage_tracking_step.jsx';
 import WelcomeWizardWooCommerceStep from './steps/woo_commerce_step.jsx';
-import WelcomeWizardPitchMSSStep from './steps/pitch_mss/step.jsx';
+import WelcomeWizardPitchMSSStep from './steps/pitch_mss_step.jsx';
 import WelcomeWizardStepLayout from './steps/step_layout.jsx';
 
 import CreateSenderSettings from './create_sender_settings.jsx';
